backend/requirements.txt
functions-framework==3.*
flask==3.*
firebase-admin==6.*
requests==2.*
python-dotenv==1.*
beautifulsoup4==4.*
pytz==2023.*
